+++ src/components/common/ScoreBanner.tsx
