+++ src/components/navigation/GDPRConsent.tsx
