+++ src/components/pages/LegalPage/index.tsx
