+++ tsconfig.json
@@ -22,7 +22,7 @@
       "@/*": ["./src/*"],
       "@/app/*": ["./src/app/*"],
       "@/locale/*": ["./src/[locale]/*"],
-      "@/components/*": ["./src/app/components/*"],
+      "@/components/*": ["./src/components/*"],
       "@/pages/*": ["./src/app/components/pages/*"],
       "@/server/*": ["./src/server/*"],
       "@/types/*": ["./types/*"],
